Using jsonrpc-macros
