+++ rpc/src/lib.rs
@@ -24,6 +24,8 @@ extern crate serde;
 extern crate serde_json;
 extern crate jsonrpc_core;
 extern crate jsonrpc_http_server;
+#[macro_use]
+extern crate jsonrpc_macros;
 
 extern crate ethcore_io as io;
 extern crate ethcore;
