+++ rpc/src/v1/impls/parity_set.rs
@@ -26,7 +26,7 @@ use fetch::{Client as FetchClient, Fetch};
 use util::{Mutex, sha3};
 
 use jsonrpc_core::Error;
-use v1::helpers::auto_args::Ready;
+use jsonrpc_macros::Ready;
 use v1::helpers::errors;
 use v1::traits::ParitySet;
 use v1::types::{Bytes, H160, H256, U256};
